Start to break up the giant class
